requirements.txt
Flask
requests
google-earth-engine
python-dotenv